+++ src/main/java/fr/aym/acsguis/component/layout/BorderedGridLayout.java
@@ -7,18 +7,17 @@ import fr.aym.acsguis.component.style.ComponentStyleManager;
  *
  * @see GridLayout
  */
-public class BorderedGridLayout extends GridLayout
-{
+public class BorderedGridLayout extends GridLayout {
     private final int paddingLeft, paddingTop;
 
     /**
-     * @param width Tile width, use -1 to use parent width
-     * @param height Tile height, use -1 to use parent height
-     * @param spacing Space between tiles, in all directions
-     * @param direction Primary direction of the alignment (direction of a "line")
+     * @param width           Tile width, use -1 to use parent width
+     * @param height          Tile height, use -1 to use parent height
+     * @param spacing         Space between tiles, in all directions
+     * @param direction       Primary direction of the alignment (direction of a "line")
      * @param elementsPerLine Number of elements on each "lines", use -1 to automatically fill the lines
-     * @param paddingLeft Left padding
-     * @param paddingTop Top padding
+     * @param paddingLeft     Left padding
+     * @param paddingTop      Top padding
      */
     public BorderedGridLayout(int width, int height, int spacing, GridDirection direction, int elementsPerLine, int paddingLeft, int paddingTop) {
         super(width, height, spacing, direction, elementsPerLine);
@@ -27,12 +26,12 @@ public class BorderedGridLayout extends GridLayout
     }
 
     @Override
-    public int getX(ComponentStyleManager target) {
-       return super.getX(target)+paddingLeft;
+    public float getX(ComponentStyleManager target) {
+        return super.getX(target) + paddingLeft;
     }
 
     @Override
-    public int getY(ComponentStyleManager target) {
-      return super.getY(target)+paddingTop;
+    public float getY(ComponentStyleManager target) {
+        return super.getY(target) + paddingTop;
     }
 }
